decompose components
