+++ src/components/Navbar/Watchlists/NavbarAssetsDropdown.module.scss
@@ -1,5 +1,5 @@
-@import '~@santiment-network/ui/variables';
-@import '~@santiment-network/ui/mixins';
+@import '../../../../node_modules/@santiment-network/ui/variables';
+@import '../../../../node_modules/@santiment-network/ui/mixins';
 
 .wrapper {
   display: flex;
