+++ src/components/Navbar/Watchlists/NavbarAssetsDropdownWatchlist.module.scss
@@ -1,5 +1,5 @@
-@import '~@santiment-network/ui/variables.scss';
-@import '~@santiment-network/ui/mixins.scss';
+@import '../../../../node_modules/@santiment-network/ui/variables';
+@import '../../../../node_modules/@santiment-network/ui/mixins';
 
 .wrapper {
   box-sizing: border-box;
